Don't hard-fail on unknown indentation levels
Non-trivial to determine context of a line. Just going to leave
indentations that are not multiple of 4 alone. Might cause some strange
behaviour in some strings, but this is how it is for now.

The feature is opt-in, so won't break anything for those that don't want
to use it.