Update black requirement to >=21.5b1
Black had a refactoring that caused some of its internals to move
around, so a small update was needed.

Intentinally keeping the `setup.py` with a `>=` requirement, because
while black is still evolving (and in beta) I want to notice these sorts
of things as soon as possible. That might not be the nicest experience
for everybody, but not the end of the world either.