normalize newlines and edit test notebook
I'd like to move towards having the same SQL formatting applied by blackbricks as what you get by using "Format SQL" in
the Databricks notebook UI. It removes any leading or trailing newlines, which this commit now ensures. It keeps your
fix for the actual bug where blackbricks would drop code on the same line as the %sql.

I've edited the test notebook too, as it is meant to hold pre-formatted code. The test checks that blackbricks does
_not_ alter this file. While what @de-hru added are valid and good test cases, something else is required to add test
with notebooks that _should_ change. Blackbricks has a longstanding need for a better testing harness, but thats a
separate issue which shouldn't be addressed as part of #32.